Google colab notebook for result to meaning
Test site, took the 2 csv files used in this project (one for the 102 classes of the oxford dataset as well as a csv of flower meanings) to figure out how to use them with each other. I did end up tweaking both csv files just a little bit on my own using a text-editor.